Backport rust dependencies changes (#11146)
* backport: chore(rust): Update web dependencies

No longer need to be patched, since they now ship it.

* backport: Bump libsodium-rs to v0.2.0

* backport: fixup! ci(rust): Run `cargo-deny` check with all features

* backport: chore(rust): Update `fuser` to 0.16.0

* chore(rust): Update backtrace to `0.3.75`

This remove the unmaintained `adler` crate

---------

+++ libparsec/crates/platform_mountpoint/src/unix/filesystem.rs
@@ -258,7 +258,7 @@ pub(crate) static LOOKUP_HOOK: Mutex<
 macro_rules! add_capabilities {
     ($config:expr, $capability:ident) => {
         $config
-            .add_capabilities(fuser::consts::$capability)
+            .add_capabilities(fuser::consts::$capability as u64)
             .expect(concat!("Capability available: ", stringify!($capability)));
     };
